xen_disk: Disable file locking for the PV disk backend
Since QEMU 2.10 (or qemu-xen-4.10), qemu locks disk images to avoid
them been re-opened in a different qemu process.

With Xen, there are two issues:
- For HVM guests, a disk image can be open twice! One by the
  emulation driver, and one by the PV backend.
- During migration, the qemu process of the newly spawned domain may
  attempt to access the disk image before the domain been migrated
  and the qemu process are been completely destroyed.

Migration of HVM guest as been taken care of in libxl, but migration
of PV guest with qdisk and HVM guest attempting to access the PV disk
before unplugging the emulated disk are still an issue.

For these reasons, we don't want to have QEMU use a locking mechanism
with the PV backend.

This is already done by db9ff46eeb in QEMU upstream, or QEMU 4.0.

Affected version of QEMU are:
- qemu-xen of Xen 4.10 and 4.11
- QEMU 2.10, 2.11, 2.12, 3.0 and 3.1

Signed-off-by: Anthony PERARD <anthony.perard@citrix.com>